backend/connect4/requirements.txt
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==3.2.13
django-channels==0.7.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
drf-yasg==1.21.7